+++ src/molgenis/bbmri_eric/eric.py
@@ -1,8 +1,9 @@
-from typing import List
+from typing import List, Optional
 
 from molgenis.bbmri_eric.bbmri_client import EricSession
 from molgenis.bbmri_eric.errors import EricError, ErrorReport, requests_error_handler
 from molgenis.bbmri_eric.model import ExternalServerNode, Node, NodeData
+from molgenis.bbmri_eric.pid_service import PidService
 from molgenis.bbmri_eric.printer import Printer
 from molgenis.bbmri_eric.publisher import Publisher
 from molgenis.bbmri_eric.stager import Stager
@@ -15,12 +16,13 @@ class Eric:
     Main class for doing operations on the ERIC directory.
     """
 
-    def __init__(self, session: EricSession):
+    def __init__(self, session: EricSession, pid_service: Optional[PidService] = None):
         """
         :param BbmriSession session: an authenticated session with an ERIC directory
         """
         self.session = session
         self.printer = Printer()
+        self.pid_service: Optional[PidService] = pid_service
 
     def stage_external_nodes(self, nodes: List[ExternalServerNode]) -> ErrorReport:
         """
@@ -49,8 +51,11 @@ class Eric:
         Parameters:
             nodes (List[Node]): The list of nodes to publish
         """
+        if not self.pid_service:
+            raise ValueError("A PID service is required to publish nodes")
+
         report = ErrorReport(nodes)
-        publisher = Publisher(self.session, self.printer)
+        publisher = Publisher(self.session, self.printer, self.pid_service)
         for node in nodes:
             self.printer.print_node_title(node)
             try:
@@ -80,34 +85,25 @@ class Eric:
     @requests_error_handler
     def _stage_node(self, node: ExternalServerNode):
         self.printer.print_sub_header(f"📥 Staging data of node {node.code}")
-        self.printer.indent()
-
-        Stager(self.session, self.printer).stage(node)
-
-        self.printer.dedent()
+        with self.printer.indentation():
+            Stager(self.session, self.printer).stage(node)
 
     def _publish_node_data(
         self, node_data: NodeData, publisher: Publisher, report: ErrorReport
     ):
         self.printer.print_sub_header(f"📤 Publishing node {node_data.node.code}")
-        self.printer.indent()
-
-        warnings = publisher.publish(node_data)
-        report.add_warnings(node_data.node, warnings)
-
-        self.printer.dedent()
+        with self.printer.indentation():
+            warnings = publisher.publish(node_data)
+            report.add_warnings(node_data.node, warnings)
 
     def _validate_node(self, node_data: NodeData, report: ErrorReport):
         self.printer.print_sub_header(
             f"🔎 Validating staging data of node {node_data.node.code}"
         )
-        self.printer.indent()
-
-        warnings = Validator(node_data, self.printer).validate()
-        if warnings:
-            report.add_warnings(node_data.node, warnings)
-
-        self.printer.dedent()
+        with self.printer.indentation():
+            warnings = Validator(node_data, self.printer).validate()
+            if warnings:
+                report.add_warnings(node_data.node, warnings)
 
     def _get_node_data(self, node: Node) -> NodeData:
         try: