+++ src/molgenis/bbmri_eric/publisher.py
@@ -1,9 +1,11 @@
-from typing import List, Set
+from typing import List
 
 from molgenis.bbmri_eric.bbmri_client import EricSession
 from molgenis.bbmri_eric.enricher import Enricher
 from molgenis.bbmri_eric.errors import EricError, EricWarning
-from molgenis.bbmri_eric.model import Node, NodeData, QualityInfo, Table
+from molgenis.bbmri_eric.model import NodeData, QualityInfo, Table, TableType
+from molgenis.bbmri_eric.pid_manager import PidManager
+from molgenis.bbmri_eric.pid_service import PidService
 from molgenis.bbmri_eric.printer import Printer
 from molgenis.client import MolgenisRequestError
 
@@ -14,9 +16,11 @@ class Publisher:
     public tables.
     """
 
-    def __init__(self, session: EricSession, printer: Printer):
+    def __init__(self, session: EricSession, printer: Printer, pid_service: PidService):
         self.session = session
         self.printer = printer
+        self.pid_service = pid_service
+        self.pid_manager = PidManager(pid_service, printer, session.url)
         self.warnings: List[EricWarning] = []
         self.quality_info: QualityInfo = session.get_quality_info()
 
@@ -26,23 +30,36 @@ class Publisher:
         copied over, the data is enriched with additional information.
         """
         self.warnings = []
-        self.printer.print(f"✏️ Enriching data of node {node_data.node.code}")
-        self.printer.indent()
-        Enricher(node_data, self.quality_info, self.printer).enrich()
-        self.printer.dedent()
+        node = node_data.node
 
-        self.printer.print(f"✉️ Copying data of node {node_data.node.code}")
-        self._copy_node_data(node_data)
+        self.printer.print(f"📦 Retrieving existing published data of node {node.code}")
+        existing_node_data = self.session.get_published_node_data(node)
+
+        self.printer.print("✏️ Preparing data")
+        with self.printer.indentation():
+            Enricher(
+                node_data, self.quality_info, self.printer, existing_node_data.biobanks
+            ).enrich()
+
+        self.printer.print("🆔 Managing PIDs")
+        with self.printer.indentation():
+            self.warnings += self.pid_manager.assign_biobank_pids(node_data.biobanks)
+            self.pid_manager.update_biobank_pids(
+                node_data.biobanks, existing_node_data.biobanks
+            )
+
+        self.printer.print("💾 Copying data to combined tables")
+        with self.printer.indentation():
+            self._copy_node_data(node_data, existing_node_data)
         return self.warnings
 
-    def _copy_node_data(self, node_data: NodeData):
+    def _copy_node_data(self, node_data: NodeData, existing_node_data: NodeData):
         """
         Copies the data of a staging area to the combined tables. This happens in two
         phases:
         1. New/existing rows are upserted in the combined tables
         2. Removed rows are deleted from the combined tables
         """
-        self.printer.indent()
         for table in node_data.import_order:
             self.printer.print(f"Upserting rows in {table.type.base_id}")
             try:
@@ -53,12 +70,14 @@ class Publisher:
         for table in reversed(node_data.import_order):
             self.printer.print(f"Deleting rows in {table.type.base_id}")
             try:
-                self._delete_rows(table, node_data.node)
+                with self.printer.indentation():
+                    self._delete_rows(
+                        table, existing_node_data.table_by_type[table.type]
+                    )
             except MolgenisRequestError as e:
                 raise EricError(f"Error deleting rows from {table.type.base_id}") from e
-        self.printer.dedent()
 
-    def _delete_rows(self, table: Table, node: Node):
+    def _delete_rows(self, table: Table, existing_table: Table):
         """
         Deletes rows from a combined table that are not present in the staging area's
         table. If a row is referenced from the quality info tables, it is not deleted
@@ -69,17 +88,23 @@ class Publisher:
         """
         # Compare the ids from staging and production to see what was deleted
         staging_ids = {row["id"] for row in table.rows}
-        production_ids = self._get_production_ids(table, node)
+        production_ids = set(existing_table.rows_by_id.keys())
         deleted_ids = production_ids.difference(staging_ids)
 
         # Remove ids that we are not allowed to delete
         undeletable_ids = self.quality_info.get_qualities(table.type).keys()
         deletable_ids = deleted_ids.difference(undeletable_ids)
 
+        # For deleted biobanks, update the handle
+        if table.type == TableType.BIOBANKS:
+            self.pid_manager.terminate_biobanks(
+                [existing_table.rows_by_id[id_]["pid"] for id_ in deletable_ids]
+            )
+
         # Actually delete the rows in the combined tables
         if deletable_ids:
             self.printer.print(
-                f"Deleting {len(deletable_ids)} rows in {table.type.base_id}"
+                f"Deleting {len(deletable_ids)} row(s) in {table.type.base_id}"
             )
             self.session.delete_list(table.type.base_id, list(deletable_ids))
 
@@ -93,17 +118,3 @@ class Publisher:
                     )
                     self.printer.print_warning(warning)
                     self.warnings.append(warning)
-
-    def _get_production_ids(self, table: Table, node: Node) -> Set[str]:
-        try:
-            rows = self.session.get(
-                table.type.base_id, batch_size=10000, attributes="id,national_node"
-            )
-        except MolgenisRequestError as e:
-            raise EricError(f"Error getting rows from {table.type.base_id}") from e
-
-        return {
-            row["id"]
-            for row in rows
-            if row.get("national_node", {}).get("id", "") == node.code
-        }